New Component: First Person Camera
This component was created to be attached to your camera to give you a
great mouselook control. Very useful for First Person games.

To use, add a camera in your scene, parent them into your character
capsule (you can use the Character controller Component on it), and
attach this Component to the camera. Don't forgot to position the camera
in a place near the "head" of your character.