New Component: Third Person Camera (+Some changes)
This component was created to be attached to your camera to give you a
great third person mouselook control. Very useful for Adventure games,
RPGs, Open Worlds, or any kind of games that may require a third person
camera.

I also made some changes on the First Person camera, so they don't
request a parent object to work anymore (useful to create a debug
camera). On the Character Component, I added a function to avoid the
character from rotating during the jump.